#209 Refactor pipeline class and add methods extractions in  build
