+++ api/gitlab/utils/error_messages.py
@@ -1,0 +1,2 @@
+NOT_FOUND = {"message": "NOT FOUND", "status_code": 404}
+UNAUTHORIZED = {"message": "UNAUTHORIZED", "status_code": 401}
